requirements.txt
tqdm==4.50.2
torchvision==0.8.2
pandas==1.1.3
torch==1.7.1
numpy==1.19.2
PyYAML==5.4.1
